requirements.txt
google-api-python-client==1.6.2
httplib2==0.10.3
oauth2client==4.0.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
rsa==3.4.2
six==1.10.0
speedtest-cli==1.0.2
uritemplate==3.0.0
wheel==0.26.0
